Fixed light ambiant effects.
This was a nasty thing to fix:
First of all, the engine was using FBO to create
overlay, but this was crash prone due to the fact
that the SDL_GL_LoadLibrary() call was missing,
and the APIENTRY macro was also missing before function pointers typedefs,
leading to possible stack corruption (mostly noticed on windows, i guess).
See the SDL doc about it.

If first tried to fix that but the fbo simply didn't load well,
since the fbo weren't working depending on the opengl version the machine
was supporting (FBO are an extension before 3.0 and are part of the core
functions after that IIRC)
Since it made all of it quite nasty to handle,
and could not be added without affecting the GUI display also,
(you can't read the text if it's also in the dark, right?)
I simply removed it, and make a blended texture based solution for that.

I also splitted the draw functions of the game modes into two, to only
make the actual game actors and scenes be affected by the ambient lights.

I also stop making the images/sprites/... use the light color directly
since it's prevented using blended colors (with opacity < 1.0f) without
making them translucent also, which wasn't the desired effect.

Finally, I removed the ApplyLightning effect function from the available lua functions,
since it should only be dealt with by the engine now.

I also removed the now useless members related to the fbo solution.
You can now add/remove ambient lights in the scripts using this, for instance:
VideoManager:EnableLightningOverlay(hoa_video.Color(0.0, 0.0, 0.0, 0.5));
VideoManager:DisableLightningOverlay();