Fixed the path finding algorithm.
Another tricky one, that will need more attention,
but already working so far.

The main culprit was the FindPath function:
The node col and row were inverted when checking the sprite collision.
Also, the function was overriding the sprite position in order to check
the sprite collision, but could return before setting the actual position
back.
Finally, the _path SpriteMovementevent member was asked as a reference,
which is poor API IMO, since you could potentially exit the function
with a random state for it.
(even if it wasn't the case, but I disliked the principle.)

I then created a function checking the collision as if the sprite
was at a given position without overriding its position.
I also changed the collision type to be more game oriented.
There is no point in distinguishing the grid collision and the map
bounds collisions for instance when the algorithm can handle both
in quite the same way.

I added a function returning the collision type corresponding
to the map object type, and documented the collision types,
according to their current use.