Finally found out where the path finding x/y tangle was.
The x/y delta, and source position were tangled in the FindPath function,
and the SetSpriteDirection checks were also inverted,
but this was done because of the destination pathnode being
tangled in the PathMoveSprite event functions.

I've untangled all that now.

The _ResolveCollision() and DetectCollision() functions
might also need more love, since they're directly or not using
checks on the grid which loaded with inverted coords.
i.e: _collision_grid[y][x] instead of _collision_grid[x][y].

I'll fix that in another commit but it's already a good step further.