Disabled the collision check for the player, when a path is being followed.
When you're playing, you don't want to see your character stuck because
of a small wall corner or an npc going in the way.

Once the path is computed, the engine should follow it,
as checks for NPCs and wall are made there already.

This isn't perfect but will avoid some rare ugly cases.