Readded lightning effect support - part 1/2
The old effect used a lit file made on windows with a custom application.
As i couldn't find any file corresponding to the need and because
the format seemed closed, I changed the way the lightning data
are loaded to use a lua script instead.

I also made the lightning effect no more create a StillImage
each frames, which should save a few cycles.

While on it, I made the boot menu use the light overlay instead
of the heavy old light overlay function.

For now, only the video effect part is done, I'll add
synchronized lightning sound support in the next commit.

As a demo, I added the lightning effect in the first map.