Fixed and readded working particle effect support.
I basically had to rework the loading function to use
the scripting engine reader.

And I had to fix several coords set mismatch as the particle
engin is currently using a inverted y coord metric compared to
the rest of the engine.

I also fixed an infinite loop in the Particle destroy function.

I also adapted and tested all the files provided  by Roos,
which I bow in front of, for his quality of work on this engine.

I added a rain particle effect on the opening map, which is looking
very nice along with the storm effects.

For now, the engine can only support static scripted particles
on screen, but it was already enough work. Thus, I'll add support
for map particle effects in another commit.