Made the engine speak in milliseconds again for animation.
It was before using a custom frame period which simply was
the milliseconds per frame / 10.
This led to a bunch of function and members used to make the
translation, and cluttered the code quite much.

I translated it all back, and added a new function loading
an animation through a lua file, permitting to easily set up
custom length per frames.

I then used it to load the battle attack animation, fixing them
on the way.

The map animation are still to be fixed but this can be done
in another commit, with the preference going on fixing
the map file and not the loading code.