Upgraded the battle effects indicators.
I also made them only update their positions when the battle
isn't paused.

Now the damages are thrown out of the actor using an actual physic
algorithm.
The heal and positive status effect are displayed next to the actor,
moving up.
The negative effects are moving down, and the miss effects
are moving through the actor horizontally.

This should make the battle look a bit better already.