Upgraded the battle character loading code to be more scriptable.
The biggest work was on adding proper sprites for the soldier
and add the proper script for all of them.

Note that I already prepared the dodge scripts ;)

I'll add the other characters after that and remove them from
the map sprites is not used there as a standard animation.