Made the engine much more tolerant with missing files.
Now the engine will report when loading a file failed,
but will no longer exit, display an ugly white rectangle,
or simply say nothing for the following cases:
- save menu
- status menu
- battle damage portraits

I also made the engine display a default stamina icon
when not finding it instead of exiting.

This all permitted me to remove a bunch of useless stub files.