Added a battle dying sequence for the battle actors.
Now the characters will triggered a dying animation and
a dead animation respectively, and fall back to an idle one
when it is not existing.

The monsters will fade away when dying.

The battles won't end before the last dying sequence has ended.

While on it, I fixed the bug in the animation location
when the battle is in command mode.