Fixed the item management between battles and inventory.
Now items used in battle will be removed from inventory
upon victory.

I had to make some API renaming to make clear what kind
of items were added and removed, and remove two dangerous
functions touching the inventory directly.

I also made the battle item list be refreshed when using an item
and prevented to use an item when none was left in a slot.

While on it, I also made to put the items reserved for use
back to battle items before committing to inventory.