Added support for save points.
Save points are system map objects gicing their position at save
time, and permitting the globalmanager to find back this position
later when loading.
It's then up to the map script to handle that, which I did
for the river access area, by adding two fully working save points.

I also removed the save option from the menu mode,
and removed also the useless load/save/cancel  option list,
since the player won't want to load while in game most of
the time. He/she will go back to the main screen if willing to load.

It's even better to not mix the save and load options, as I found
myself load instead of saving once, which is very frustrating.