Made the ambient, light and lightning effect handled per game mode.
For that, I introduced the effect supervisor which is a component
of the mode manager.

The video manager is a bit less a god object now, too.

To function properly, I still have to make the particle manager
another component of the mode manager.

The evil plan is to sooner or later also make the cinematics
used in the boot menu and in the battle animation part
be another mode manager component.