Made the particle management be done per game mode.
To acheive that, I made the particle manager be a component
of the Game Mode.

The VideoManager is now even less a god object about that point.

I also optimized the particle manager by merging the API
about loading the particle effects, and made it use a vector
to store the loaded effects which can't hurt performance.

Conflicts:

	src/engine/video/video.h