The big editor overhaul - part 1.
The editor got support for an arbitrary number of layer as for the internal code.

I also added a tree view where you can select the layer you want to edit.

I had to disable the map resizing, a part of the autotiling feature for now,
and the possibility to set a lyer's visibility,
so i could reach a convenient enough base to commit.

Note also that I intentionally removed the object layer support,
as it seemed incomplete, and not really necessary, at least for now.