Worked around battle deaths and field revivals.
Now the characters party won't run if one of them is dead
at battles starts and ends.
A new Elixir item can revive a character which has 0 HP.

Made sure also that the potions can't revive a character,
and the MoonJuice Potion can't raise SP of a character in coma.
When the use has failed, the potion is currently still removed, yet.