Made every events linked to a sprite be paused when you're speaking to it.
This way, even delayed events will be paused,
preventing strange random behaviour.

I also added two functions checking for every sprite event function
instead of an event id, since one can't pause delayed event with
the id of its parent. Plus, this is not what was intended anyway.

I adapted the PauseEvents and ResumeEvents functions for now
but I might completely remove them later as they are useless for now.

This is fixing the dad and mother behaviour when you're speaking to them
while they're not moving (as they are waiting for the next delayed event
at that time.)