Finished the context inheritance support.
Previously, parts of the editor were made to support another
inheritance than the from the base context, but the game
didn't support that.

I added support for inheritance from a previoulsy loaded context
(or none), and made the first context nameable along the way.

The maps still need to be translated but that will be part
of another commit.