Merged back the map objects position and offset into one float.
This permitted to simplify a bit the overall API,
remove the need to check offsets > 1.0f, and made possible
to precise the path move event destination node, which is
a first step toward making it fully work.

The detection algorithm has still issues with wall collision
but the fix is still being studied.

Note that I kept on purpose the pixel rounding to keep avoiding
the tiles flipping, which was the main cause of splitting
the position from their offsets, even if I appeared to be not quite
necessary.