Rewrote the collision detection algorithm.
The previous implementation relied on the object supervisor,
and was done using several very complex functions, that seemed
to not do what they were intended for.

So I preferred to rewrite it all in one function based on the sprite,
aimaing at dealing at each case sequencially.
And it happened to get out quite good so far.

This change permitted to:
- Make the player, and npc fall back to a straight direction
when walking diagonally and a collision was detected in only
of the two active directions.
- Make the path move event work as expected, by making the npc,
camera ignore the collisions when following a path, but:
- Making two tangled npc/camera (with interleaving collision rectangle)
being free to get out of each others.

IMO, this will take care of 95% of the cases, and even with playing
a lot with it, I couldn't find no-op and stucking cases so far.

walking in the maps is a lot more smooth now :)