Finally found out how to remove the white edges often seen on sprites.
The artifacts seen where coming from the GL_LINEAR algorithm
using the RGB values of the transparent pixels to setup the average
pixel value when smoothing.

To fix that, I simply reset back the rgb values of transparent pixels
to pure black.

The fix was simple but this issue is officially the one that took
me the longest to figure out.