Changed the smooth video options to only change the map tiles rendering method.
I'm still not convinced about that code, and I'm quite wel aware of the graphical
glitches seen in the map mode.
This specific flaw will have to be thought very carefully.

Yet, rendering each and every textures not smoothed was silly since
the presentation can't look right without any smoothing method.
In conclusion, not good but a small step in the right direction.