Rewrote the map sprite loading system to use completely scriptable definitions.
I also cleaned up some code along the way, and turned private a useless? member
in the MapEnemy class.

Now the Map Sprites can use custom and scriptable delays for each frames,
and can use an arbitrary number of frames in an arbitrary order,
and an arbitrary time. This was a long awaited step needed to add
true custom animation support for instance.

The characters' speed is yet to be turned into a speed factor
for the sprites motion, as now the sprites won't seem to accelerate
when their speed is increased, but it's a minor point that
can wait a few commits.