Turned the welcome window into a help window.
I made it available to every game mode when pushing F1 key.
Also, I made the text displayed be dependent of the current
game mode, making it possible to have in-game help!

On the low side of things:
- I didn't finish all the texts for now, so certain game modes
still have a generic one.
- The text isn't scriptable and the positions are hard-coded.
(I guess it'll stay like that for a while.)
- That changes showed up that the editor is heavily depending
on the VideoManager because of simply using a StillImage
in the Tileset Editor. Something I'll need to take care of.

Hence, I'll polish all that in other commits.