Added support for scriptable particles effects and particle map objects.
To achieve this, I had to upgrade the particle manager to give back
ParticleEffect pointer instead of id, and add several functions
to the luabind interface.

I also created a ParticleObject class which is created a particle
effect without registering it to the (rather inefficient) particle
manager, which permitted me to draw the effect at the right time,
i.e: depending on the map objects y coord order.

I also fixed the effect offset handling for circle shaped effects,
and added active and inactive particle effects for save points,
which you can now see in the forest entrance.