Strengthen the battle target choice logic.
Overall, I returned when necessary instead of letting the engine
in an uncertain state sometimes.

I also added an extra check on actor HitPoints, since their state
might not suffice to compute whether they're valid or not.
I also forbid the choice of dead enemy targets in all cases,
since it won't be seen in the game play.