Finished support of Active, Semi-active and wait type battles.
To acheive this, I first removed the mess about state timers management
between the battle actor and battle class and simply made the battle
instance tells all the actor when to update their state.

Using this, I turned the battle type into a battle setting instead
of a game one, permitting to introduce different kind of battles
throught the game.

Now:
- Wait battles (used at game start) are battles where activating
a character command will pause the battle, but not in progress animations
(idles ones will play, and attack will finish in that case, for instance.)
The idle state of players and enemies is in that case reduced by 3,
making the battle faster and thus less boring since all the decisions can be taken
with all the time needed.

- Semi active battles are battles where activating a command will prevent
all the actors to attack but not to complete their idle and warmup states.
Meaning that you can stop the monsters from attacking you while thinking
but only to a certain point.
The battle idle time is divided by 1.5, a little bit faster than the normal
speed.

- Active battles are battles were nothing stops, and you have to hurry
chaining commands to win.
The battle actor idle time speed is normal in that case.