Strenghten again the target selection logic in battle.
Now the logic will search for the next target in the list if the previous one
failed, if that fails also because it can if the target was invalidated
within the same battle loop, if will then search for a completely
new target.

This fixes the last (and quite rare) case where one character
can't select an attack target anymore.