Made the battle weapon ammo independent in term of draw sorting.
I added the BattleObject class making all the visuals depend on it.

It permitted to make the ammo be independent in term of draw sorting,
but will also permit to later add proper battle particle effects for instance.

Added the BattleObject class and let the actor and ammo be yet another type of object