Upgraded the fonts used in Valyria Tear.
The commit is based on the complain of several people telling
it was kinda hard to read the text sometimes, especially the numbers.

By using bold fonts, and a sans serif one for the numbers
actually helps a lot in the readability without changing the space
used.

I didn't changed the font internal name to reduce the commit scope,
and I'll come back to that by making it more flexible in another commit.

Changing the font wasn't an easy choice as it quite change the personality
of the overall game.