Improved a bit the equipment handling in the menu mode.
First of all, I turned PHYS ATK/DEF into ATK/DEF alone,
META ATK/DEF into M.ATK/M.DEF, as this is shorter
and many people are used to those names out there.

As for the menu actually, I made only the equipable items
appear when trying to equip, and made the stats diff appear
bottom right. I.e.: If you try to equip a stronger armor,
it will show a +X in green bottom right, and a -X in red
if it is weaker.

The menu drawing logic is still very unoptimized, but this kind
of fix belongs to another commit.