Thanks to previous commit, hopefully fixed the script invalid pointer detection.
the save menu was opening the map file where one had previously saved,
this caused no troubles from the boot menu, but was creating an invalid
opened file when looking at the save menu from within the map file
one had earlier saved in.

To fix that, I no longer permitted to open the same file twice
(which is bad anyway), and rearranged the code to make sure
it doesn't happen.