Profiling: Turned the font glyph cache into a vector.
The fact of using a std::map for the glyph cache made the application
spend 12% of the time do find() calls on my own.

Turning this into a vector made this kind of calls disappear
from the profiling top list and should save a lot cycles.

The application should run a bit smoother now.