Fixed crashes caused by errors in the status script.
There was a typo in the evade raise status effect leading to a crash,
and a obsolete statement in the paralysis one, leading to a crash
with the Stun Strike.

I had to remove the paralysis effect of the stun strike since
the necessary code isn't ready yet to add that effect.