Implemented the stun effect.
Also mde the status effects timers update only when the battle isn't
waiting for the player to be a little bit more fair.

By avoiding to use the StunTimer() function, I also was able
to get rid of the BattleTimer class which look useless to me,
and corner case-prone.

I also added a 'Zzz' icon displayed on top of the enemies
when they're stunned.

Note also that the stun effect can miss, separately from the
damage chances.