Made the elemental and status effects loaded in equipment, and displayed in the shop mode.
I also renamed the sockets -> shard slots, removed the lore aka long description
which is not used and won't likely be.

I also turned the elemental and status effect std::map into vector
to speed up the display time a bit.

I then added the Paladin Sword in the debug shop to show
the elements and status effects there.