Turned the use of script object back to new allocated pointers.
This was partly the former use of it, and it took me time to figure out
why.

Some crashes at app exit were thus caused by my changes, but also
by other code part. This should lessen the crashes seen at app exit
a lot, even if luabind can still cause a crash in lua for a reason
I'm not aware of.