Finally ... Fixed certain tiles not correctly drawn in the map mode.
Found an off-by-one offset in the glViewport() calls.
I also reduced the glTranslate() calls offset to make it more compliant
with the actual pixel size in that mode.

I might not be perfect but at least the floor problem seen in the video
is gone both on Windows and on Linux, as for me, at least.

Closes #2