bug fix and slight item menu change
bug fix:
+ when using an item and causing that item category to "run out", a second
press of the use button (default f) would cause the game to crash.
As described in #99

this fix also addresses another issue by changing (slightly) how the items
menu operates:

+ When an item is used or equiped, we return back to the item list. When a
list is empty due to the previously mentioned "use/equip", we now return
back up to the category selection. If there are no more items in the
inventory, we are returned all the way back up to the Use/Equip/Remove
selection stage. In essence, it acts as a "stack unwind."

This fixes the bug related to this issue, as well as an ticketed bug/issue
whereby once an item ran out, the list automatically moved to the next
item in the list without warning. this meant that you could accidently use
up items with rapid button presses and not have a "one-button" cushion.