Small code fix/cleanup after Ikarus' work.
Mostly, I added the license info about the test map,
turned the mutable reference into a pointer, not that it was bad,
rather because it is the current use in the rest of the code,
and made the location names translated.

We might want them to be ustring btw.
I'll also add new location marker in another commit.