Removed the Swap, left select and right select keys.
Those keys were not obvious in use, and I do think there are
already enough keys so far.

That's why I replaced the use of the swap key by the menu key where possible
and defuncted the swap key.

The left and right select keys were not really working in certain case
and are very rare to use so I removed them as well.

This should also add some place to permit adding the missing joystick
keys and features.