Turned the ReloadSprite() function back to a c++ one.
Linked to #112:
As shirish reported having 'virtual function called' based segfault,
and as the bug happens within the function that reload the sprite,
I've suspected the map sprite reference badly taken has a map object
one, or something similar.

If it's not the case, at least this will be a cause less to see that
bug, or it will throw the error from within a c++ function, which
is more talkative with gdb.

A more general global variable hunt is also to be made also to
possibly get rid of that bug as well as the others.