Turned Attack and Support into Weapon and Magic! - Part 1.
This makes the later mechanics much more logical as the plan
is to make certain skills available only through the equipment
or shards.

While on it, I also added icon support to skills, and added icon
the magic skills viewable in the menu mode.

I also fixed Sylve starting/debug equipment.

The battle mode is left to be adapted.