Better balanced the enemies stats.
I took the time to create a sheet where I can have a look at the XP curve
of both the characters and have a better way to define the enemies XP
and Drunes based on their stats, objects given and their specialties.

I then added a bit more XP on first comers, and fixed a mistake in
the Slime Mother's and second Fenrir HP.

I need to test it out but it should be a bit better.