Split of map data and script - Step 1
Adapted the map and save code so that it's possible to load the script
part from a another script file, while still permitting to load it
from the same file.
This permitted to remove the duplicate crystal map data from the introduction
script.

The next step will be to separate the script part of the map having no savepoint
as it will remain backward compatible.

No change should happen as for the user.