Finally removed the last game compilation warning.
The scroll offset parameter was always equal to the same named member
so it was useless in any case.

Secondly, horizontal and vertical scrolling triggers a movement,
so it is basically working.
Yet, the scrolling movement is always from top to bottom, so other
scrolling movement are to be added but the scroll offset is also
useless here.