Fixed the crash happening when quitting the game during a transition to battle.
While digging around something I'll push in the next commits,
I found that the BattleMode singleton instance was set in the reset() function
leading to the fact that when it was freed before being initialized and reset,
the finish supervisor would make the game crash due to calls to that instance.

I made sure the transition to battle would free the battle mode if it was destructed
before passing to the game mode manager and fixed that crash by setting the instance
in the battle mode constructor, as no more than one battle at a time will happen anyway.

Sorry Ikarus, I know you wanted this one but since I saw it, I fixed it.
Hope you don't mind ;)