Merged big parts of the GmaeModeMedia class into a GlobalMedia one.
This will load the most common files used in the different game modes
instead of loading/unloading them when changing the game mode.

I also made a few misc code cleanups while doing it.

I didn't completely removed the GameModeMedia classes as certain
ressources there were still specific to their game mode, such as
the battle background, music and so on.

I also simplified the way status effect icons are queried, avoid
the use of an extra map for it.

I didn't touch the menu mode yet, as I'm not sure of what will change
for now. I did that before playing with the menu mode as it will
much easier to query status effect icons now :)

But this is IMHO sufficient to close #121.