Equipment info and menu improvements - Step 1
- Moved the Drunes and time info onto the MenuMode class.
It was duplcated in each AbtractState class children before that.
- Merged the update and the drawing of equipment info
and show status, elements and shard slots properly.
- Optimized the equipment update and drawing logic. The menu will now
update the equipment info only when necessary.
- Made several other adjustements.

What's left is showing the diff with current character equipment part,
and properly show the equipment stat when using the items menu.

I prefered splitting this change in another commit as this one is big
enough as is.