Dropped context support - game part.
The drop of context support will permit later to use Tiled as a map editor,
and permitted several other simplifications and last but not least lighten drastically
the map layer draw code in term of cycles needed to do it.

The next part will be to adapt the editor and get rid of the obsolete data in the map data files.