Removed the tile smoothing options from the settings file.
I may readd it later but since players from the previous game version
had the tile not smoothed per default due to my own stupidity.

This will permit to let everyone try the smooth option in the next version
as very few were actually aware of this 'bug' or feature depending on
how you see it.