Introduced battle scene mode support.
This permit to arbitrarily pause the battle, not only for dialogues
but also other future scripted events.

It also fixes a bug in the battle effects which were updated twice,
once manually, once because the timer is handled by the system manager.

The status effect couldn't also be correctly paused either before.