Turned actor stats into a container handling a base + a modifier.
This will permit to never touch the base value and simply apply modifiers
on it, just as it is done with the attack points.

For now, the active (from battle) modifiers are applied correctly.
The code applying the passive ones (still in the battle code) is in fact wrong
and should be moved to the globalactor class along with a good fix.
Yet, this will be for another commit.

Along the way, this was also done and/or needed to be done:
- Removed dead code
- Fixed the enemies using the strength value as vigor in battles.
- Ironed out the status effect lua function calls to prevent game crashes,
  seen mostly in the debug battle.
- Fixed bugs in the status effect Remove() functions and old enums used
  in the elixirs lua code.