Big status effect handling overhaul.
- Merged elemental effects into status effects completely.
I removed them from being applied within the equipment as
it was the wrong way of doing it.

- Added support for passive and active status effects in battles,
and when equipping/unequiping.
I thus removed the old status effect handling code from the battle mode.

- Simplified the status.lua file by removing duplicate scripting,
and added support for passive effect calls there and in the engine.

- Fixed the improper use of copy construction for global stats instead
of setting their base value.

- Added support for SP damage/stealing to finish an effect.

The status effect code is now much saner and extensible. :D
What's lacking now is some code to call updates for passive (equipment) effects
in battles and proper support of both passive and active effects in the map mode,
but at least the core mechanisms are in a good shape now.