Made the passive effects be updated in battles.
I also started moving some common logic back to the global status effect
class.

And I tried to clarify in the status.lua what is a battle function
and what is global one.

What's left is to create a passive status effect class for battles
as the use of full battle status effects is useless for passive effects.

And had the same functionalities for the map mode.

Yet, this is sufficient to close #141