Added pathfinding and way point support to enemy sprites.
Enemy sprites will now use path finding to walk and locate the character.

This feature is for me working so far with one rough edge I need to deal with soon:

When behind an obstacle, the enemy sprite will make heavy computation to try and
catch you uselessly. I need then to be able to specify an optional limit in the close list
of path nodes. I'll do that in another commit.