Added primary support for joystick hat events.
This is far from being perfect or even usable in the map mode, for instance.
Bu you can at least navigate in the menu using the XBox 360 D-Pad for instance.

I might be forced to use specialized members for that one and I'll do that
in another commit.

Part of #257.