Made TextBoxes use pre-rendered text images when using instant rendering & other optimizations.
I also made the menu windows managed through a std::vector instead of a map in the GUI manager,
and removed the dead hide/show rendering and scissoring logic in this class.

I finally also removed the _initialized parameter in GUI objects
and made them always ready for use in every possible construction, making this useless.