As a proof of concept, made the script supervisor able to create and return TextImages.
For now, I adapted the in game keys script seen at game start.
As I earned an average of 20 FPS there, I do think it's on the right track. :)

Another tested change, is that now, the script supervisor will return be turned
to return Image objects instead of custom index of them.
The Draw() method used is now the one of the object and not a wrapper in the script supervisor.

As a first improvement, this is now faster and the engine is now able to render text in map mode
through scripting using custom text styles.

This is part of the effort described in #252 and #256.