Made the scripting engine create and return images objects that can be handled through scripting.
The life cycle of the images are still handled by the engine for better memory management purpose.

The useless Draw wrappers have been removed and performance should have improved a bit thanks to this.
The scripting process of images is also more in sync with how it is done in the engine itself.

Following that API change, I had and did adapt all the scripts that created images.

this closes #256