Fixed A Lot Of Horrible English Grammar
Let's face it, the English in this game is as bad as any other Engrish attempt at English. A lot of very basic principles in grammar have been ignored. I've fixed a lot of lines, but there is still improvements that can be made. A note to make for the author(s) is that you need to make sure you are properly writing down the context in each scene. There's a lot of places that would be significantly better if the plot was more fleshed out and the spoken words and descriptions had more context. If you get stuck when writing a story, try to write the story backwards.