Fixed item handling + added a basic item and skills help window.
I had to rewrote the mechanism between the battle item and the global items
to make sure everything was working good.

As a bonus, and while on it, I restored the information window for skills
and items that you can toggle with the Menu key ('S').
This version is still basic but already very functional. :)

Closes #282.