Improved the pixel length computation to not be hardcoded anymore but computed only once.
The pixel length was actually different for each resolution so I made it computed
at first map mode update call.

The pixel length is needed for this mode to avoid seeing sprite vibrating and such,
due to the coordinate system used.

Unfortunately, it's not perfect, and I'll have to change the map mode to use
the standard coord sys to fix this the proper way.