Consolidated _any_key_press
The issue occurred because:
After _any_key_press was correctly set to true on a hat event,
it was often being reverted to false by the logic in _JoystickEventHandler during the event-handling loop.

Since _any_key_press is only assigned in the event handling loop,
and not evaluated, it makes more sense to assign it once after all events have been handled.

I tested this change in all menus that use InputEngine::AnyKeyPress(),
using the keyboard, and Xbox 360 controller joysticks and D-pad.
All seem to work as expected.