Adapt French translation to the Transifex line endings
Since Transifex uses a slightly different convention for visual text formatting,
the French translation has been upload to Transifex and downloaded back to adapt
its formatting. The content was not changed, but subsequent diffs will be clearer.

Note that Tx breaks the file description header (with the name and copyright),
so it has to be restore manually each time before pushing to git.