Big map object API overhaul - Part 1
NOTE: The map code will have to be upgraded for this to be functional.

This API change makes map objects auto-register to the object supervisor
and set their object ID.

The new API also permits to create a map object without having to add it
to a layer in a second statement.
The object can only be declared with a layer, and is owned by the engine,
making the scripting API more streamlined.

The same will be done with Map Zones, dialogues and events eventually.

Part of #410