Split the pixel, screen and tile values for collision and sprite sizes.
This way, I can start experiment on different zoom ratios, and fix the game so
I can kill the most magic number possible.

Images size and collisions are now set in pixels to fit what an artist would expect
when adding new sprites and the game auto translate the values into screen and tile based values.

At the moment, collision are still in need of a fix and particle object don't resize themselves.
Apart from that, the game is starting to be working again.