Finally fixed the memleak notably seen when changing map.
I think this was the worst Valyria Tear/Hero of Allacrost bug ever.

The bug was due to the fact that lua "thread" which are coroutine and not actual
threads were never freed from the global stack, making each script loading bloat the memory.

Once that bug was fixed, I encountered related hidden bugs:
- The audio buffers, now correctly freed, would cause an occasional segfault when loading the same music
  several times due to an OpenAL bug (Fixed in previous commit.)
- The luabind::object (AKA ScriptObject) function "pointers" would sometimes be freed after the lua coroutine
  if the garbage collector was fast enough, triggering random crashes when freeing their corresponding handle.

This is the biggest part of #480