Started implementation of battle formations. (Disabled for now)
This is a warmup to touch the menu windows before the addition of the skill tree. ;)

For now, this only adds the basic maths and interface to set up the battle formation
of characters.

What's left is to apply both visually and logically in battles.
This is also to be saved in the save games.

Part of #468