Added core support for "home" maps.
The home map is where you teleport back when you will use an escape smoke
thanks to Kalya.

Next step is to add support for teleport to the latest home map when Kalya
is in the party and you have a home map set, with proper GUI, graphical effect, ...

Once this is done, going back to the village for more equipment will make more sense.

Part of #391